add runtime tensor checks back to t3k unit test
